requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
alembic==1.16.4
attrs==25.3.0
audioop-lts==0.2.2
discord.py==2.6.0
frozenlist==1.7.0
greenlet==3.2.4
idna==3.10
Mako==1.3.10
MarkupSafe==3.0.2
multidict==6.6.4
propcache==0.3.2
psycopg==3.2.9
psycopg2-binary==2.9.10
python-dotenv==1.1.1
setuptools==78.1.1
SQLAlchemy==2.0.43
typing_extensions==4.14.1
wheel==0.45.1
yarl==1.20.1
